Move FileSystemJobState into storage dir
