+++ src/framework/execution/types/index.ts
@@ -5,3 +5,4 @@ export * from './validation';
 export * from './instance';
 export * from './logger';
 export * from './partialDatasets';
+export * from './jobState';
